requirements.txt
Django==1.11.4
django-filter==2.0.0.dev1
django-widget-tweaks==1.4.1
image==1.5.17
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
pytz==2017.2
